+++ bindings/dart/lib/src/ffi/bindings.dart
@@ -2383,6 +2383,7 @@ class QuiverDatabaseBindings {
     ffi.Pointer<ffi.Char> date_time,
     ffi.Pointer<ffi.Int> out_data_type,
     ffi.Pointer<ffi.Pointer<ffi.Void>> out_values,
+    ffi.Pointer<ffi.Pointer<ffi.Uint8>> out_mask,
     ffi.Pointer<ffi.Size> out_count,
   ) {
     return _quiver_database_read_time_series_row(
@@ -2393,6 +2394,7 @@ class QuiverDatabaseBindings {
       date_time,
       out_data_type,
       out_values,
+      out_mask,
       out_count,
     );
   }
@@ -2408,6 +2410,7 @@ class QuiverDatabaseBindings {
             ffi.Pointer<ffi.Char>,
             ffi.Pointer<ffi.Int>,
             ffi.Pointer<ffi.Pointer<ffi.Void>>,
+            ffi.Pointer<ffi.Pointer<ffi.Uint8>>,
             ffi.Pointer<ffi.Size>,
           )
         >
@@ -2422,6 +2425,7 @@ class QuiverDatabaseBindings {
           ffi.Pointer<ffi.Char>,
           ffi.Pointer<ffi.Int>,
           ffi.Pointer<ffi.Pointer<ffi.Void>>,
+          ffi.Pointer<ffi.Pointer<ffi.Uint8>>,
           ffi.Pointer<ffi.Size>,
         )
       >();