+++ bindings/python/src/quiverdb/_c_api.py
@@ -304,7 +304,8 @@ ffi.cdef("""
 
     quiver_error_t quiver_database_read_time_series_row(quiver_database_t* db,
         const char* collection, const char* group, const char* attribute,
-        const char* date_time, int* out_data_type, void** out_values, size_t* out_count);
+        const char* date_time, int* out_data_type, void** out_values, uint8_t** out_mask,
+        size_t* out_count);
 
     quiver_error_t quiver_database_update_time_series_group(quiver_database_t* db,
         const char* collection, const char* group, int64_t id,